requirements.txt
disnake==2.5.2
packaging==21.3
python-dotenv==0.20.0
replicate==0.0.1a15
apscheduler==3.9.1